+++ Script/validation_signup.js
@@ -1,11 +1,4 @@
-export const USER_DATA = [
-  { email: "codeit1@codeit.com", password: "codeit101!" },
-  { email: "codeit2@codeit.com", password: "codeit202!" },
-  { email: "codeit3@codeit.com", password: "codeit303!" },
-  { email: "codeit4@codeit.com", password: "codeit404!" },
-  { email: "codeit5@codeit.com", password: "codeit505!" },
-  { email: "codeit6@codeit.com", password: "codeit606!" },
-];
+import { USER_DATA } from "./data.js";
 
 document.getElementById("Email").addEventListener("input", ValidMail);
 document.getElementById("Password").addEventListener("input", ValidPassword);
