+++ login.html
@@ -28,10 +28,13 @@
                 </article>
                 <article class="password_box_set">
                     <label for="Password" class="password_box">비밀번호</label>
-                    <input id="Password" type="password" placeholder="비밀번호를 입력해주세요" />
+                    <div>
+                        <input id="Password" type="password" placeholder="비밀번호를 입력해주세요" />
+                        <img src="TheEye.png" alt="Eye_for_password" ID="togglePassword" class="toggleEye">
+                    </div>
                 </article>
 
-                <button type="button" class="login_button" disabled>로그인</button>
+                <button type="button" class="login_button">로그인</button>
             </div>
             <div class="simple_login">
                 <div class="simple_login_text">
@@ -48,15 +51,15 @@
                 판다마켓이 처음이신가요?
                 <a href="/signup.html" style="color: #3182F6">회원가입</a>
             </div>
-            <div class="modal" id="customModal">
-                <div class="modal_popup">
-                    <span id="modalMessage"></span>
-                    <button type="button" class="close_btn" id="closeModalBtn">닫기</button>
-                </div>
+                    <div class="modal" id="customModal">
+            <div class="modal_popup">
+                <span id="modalMessage"></span>
+                <button type="button" class="close_btn" id="closeModalBtn">닫기</button>
             </div>
+        </div>
         </form>
     </section>
-    <script type="module" src="Script/login_js.js"></script>
+    
 </body>
-
+<script type="module" src="Script/login_js.js"></script>
 </html>
