+++ style_login.css
@@ -7,7 +7,7 @@
 
 
 html {
-    font-family: 'Pretendard', sans-serif;
+    font-family: 'Pretendard', sans-serif !important;
     font-size: 16px;
     font-weight: 400;
 }
@@ -71,10 +71,26 @@ section {
 #Password {
     background-color: #F3F4F6;
     border: none;
-    background-image: url('TheEye.png');
-    background-repeat: no-repeat;
-    background-size: 24px 24px;
-    background-position: right 16px bottom 16px;
+}
+
+#togglePassword{
+    width:24px;
+    position: relative;
+    left:92.5%;
+    bottom:40px;
+}
+
+
+.toggleEye {
+
+}
+
+.visible {
+    display:block
+}
+
+.hidden {
+    display:none;
 }
 
 #Password:focus {
